Add generated Config.in files.
It seems that these are now supposed to be generated from Config.src in order to copy some extra config parameters from source code. I don't know why such a thing is desirable, but there it is. In any case, the kconfig I use currently can't handle generated config files, so I'm just checking these in as-is. In the future it would be nice to support that (or do the sane thing and move config stuff to config files, and keep the source code in the source files...)